[mod_wstunnel] fix ping-interval for big-endian (fixes #2944)
fix wstunnel.ping-interval for big-endian architectures

(thx ste_p_james)

x-ref:
  "[patch] wstunnel.ping-interval ineffective"
  https://redmine.lighttpd.net/issues/2944

+++ src/mod_wstunnel.c
@@ -107,7 +107,7 @@ typedef struct {
     gw_plugin_config gw;
     buffer *frame_type;
     array *origins;
-    unsigned int ping_interval;
+    unsigned short int ping_interval;
 } plugin_config;
 
 typedef struct plugin_data {
